migrate setup.py to pyproject.toml
