--- Pipfile
@@ -1,20 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-influxdb-client = "*"
-openant = {editable = true, path = "."}
-
-[dev-packages]
-ipython = "*"
-pytest = "*"
-pylint = "*"
-sphinx = "*"
-furo = "*"
-sphinx-mdinclude = "*"
-twine = "*"
-
-[requires]
-python_version = "3"
