+++ README.md
@@ -19,11 +19,10 @@ A note on ANT/ANT-FS/ANT+: this module is for development and testing of devices
 ## Requirements
 
 * Python >= 3.7
-* libusb 1.0 (for pyusb)
 
-Run `pip install openant` or `pip install git+https://github.com/Tigge/openant#egg=openant` for HEAD. A 'Pipfile' is also provided for use with `pipenv`.
+Run `pip install openant` or `pip install git+https://github.com/Tigge/openant#egg=openant` for HEAD.
 
-If using on Linux, a udev rule for the Dynastream ANTUSB stick can be installed with `sudo python setup.py udev_rules`. Windows does not use udev_rules and therefore does not need to be installed. Follow libusb's driver installation [instructions](https://github.com/libusb/libusb/wiki/Windows#Driver_Installation) for Windows. macOS should work with just libusb installed.
+If using on Linux, a udev rule for the Dynastream ANTUSB stick can be installed with `sudo python -m openant.udev_rules`. macOS/Windows does not use udev_rules and therefore does not need to be installed. Follow libusb's driver installation [instructions](https://github.com/libusb/libusb/wiki/Windows#Driver_Installation) for Windows. macOS should work with just libusb installed.
 
 ### ANT USB Stick
